Fix use of __ outside of vapors library
