+++ core/src/main/scala/com/rallyhealth/vapors/factfilter/dsl/InfixOps.scala
@@ -1,0 +1,3 @@
+package com.rallyhealth.vapors.factfilter.dsl
+
+final class InfixOps private[dsl] extends ComparisonOps with TypedFactOps
